(Implement) select new monster
