+++ src/components/monster/LevelOneMonstersDisplay.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
@@ -23,16 +23,12 @@ const LevelOneMonstersDisplay = ({ go }) => {
   const monsterList = useRecoilValue(babyMonsterState);
   const setSelectedMonster = useSetRecoilState(selectedMonsterState);
   const [selectedAvatar, setSelectedAvatar] = useState(() => monsterList[0]);
+  const excludeMonsterId = location?.state?.levelOneId ?? -1;
 
   const handleSelectMonster = () => {
     setSelectedMonster(selectedAvatar);
     go();
   };
-
-  useEffect(() => {
-    console.log(location.pathname);
-    console.log(location.state);
-  }, [location]);
   console.log(monsterList);
 
   return (
@@ -55,16 +51,20 @@ const LevelOneMonstersDisplay = ({ go }) => {
         <SelectList>
           {monsterList.map((monster) => {
             return (
-              <SelectListItem
-                key={monster.monsterId}
-                selected={selectedAvatar.monsterImage === monster.monsterImage}
-                onClick={() => setSelectedAvatar(monster)}
-              >
-                <MonsterThumbnail
-                  imageUrl={monster.monsterImage}
-                  imageAlt={monster.monsterImage}
-                />
-              </SelectListItem>
+              excludeMonsterId !== monster.monsterId && (
+                <SelectListItem
+                  key={monster.monsterId}
+                  selected={
+                    selectedAvatar.monsterImage === monster.monsterImage
+                  }
+                  onClick={() => setSelectedAvatar(monster)}
+                >
+                  <MonsterThumbnail
+                    imageUrl={monster.monsterImage}
+                    imageAlt={monster.monsterImage}
+                  />
+                </SelectListItem>
+              )
             );
           })}
         </SelectList>
