+++ src/components/myPage/index.js
@@ -1,6 +1,6 @@
 import UserInformation from './UserInformation';
 import Notice from './Notice';
 import EditBox from './EditBox';
-import FollowData from './FollowData';
+import FollowListItem from './FollowListItem';
 
-export { UserInformation, Notice, EditBox, FollowData };
+export { UserInformation, Notice, EditBox, FollowListItem };
