+++ src/constants/monster.js
@@ -1,1 +1,2 @@
 export const MAX_LEVEL = 5;
+export const MAX_EXP = 100;
